A number of small changes for English readability
There were some instances of incorrect English, which I corrected. In other cases, I made changes so that the text read better, cleaned up duplicate information, or re-worded the text to sound more natural and straightforward. Lastly, I took a look at the document at a high level to try and align formatting and usage so the document fits together. There is a bit more that could be done; the unnecessary use of commas, in particular, is something that I'd like to change. For me, this is a test to see if my changes are approved and accepted. If so, I'd be happy to continue reviewing the remaining KiCad documentation in a similar manner.