One correction and more updates
This is overall a smaller patch, but touching a number of areas of the document. I made the correction you noted and also re-read it all with fresh eyes which revealed a couple more minor issues and a number of areas I felt could be improved. Please review for any issues.