Fix in issue #226
Created a new section for an edited version of Kinichiro's text in the issue body. I feel that this isn't part of section 2 because one can use KiCad without accelerators or hotkeys; we all agree it needs to go in the document before we start talking about how to use the individual tools. **Need to decide on the exact syntax for hotkeys then update the rest of the document!**