Workflow and annotation sections review
I'm not keen on the sentence "Some do not consider this particular feature especially useful." in regards to Backward Annotation, but I'm leaving it for now. Also retitled the section under Using KiCad since it was just repeating the following two sections and the term "shortcut" should be more obvious to a new KiCad user. It's weird: I feel that I already submitted a patch for this...