README.adoc: additional information for Cyrillic PDF file
The README lacks a hint for building the documentation for languages
with Cyrillic characters. It took me a while to sort this out which font
package is needed for that. The Travis controling file already contains
this needed package as a dependency. Adding the needed package to the
hint about the package that are needed in Debian and Ubuntu.

This commit also is sorts the package that are required on a Debian or
Ubuntu system and also add some line breaks to make the list more C&P
friendly.