Cleanup cmake rules (#566)
* Split custom_target into custom_target and custom_command, avoid rebuilds

As stated in the add_custom_target documentation, the target "is always
considered out of date". Instead, only add the output file as a target
dependency and create the file using add_custom_command. With this change
the documentation is only recreated when the input changes.

Signed-off-by: Stefan Brüns <stefan.bruens@rwth-aachen.de>

* Remove intermediate targets and use file dependencies instead

Instead of creating custom target for each intermediate output, like
translated asciidocs, just add/append the generated output files as a
dependency.

Signed-off-by: Stefan Brüns <stefan.bruens@rwth-aachen.de>