Revert "Revert "Deleted github macros that seems to interfere with the compilation process""
This reverts commit a3c68764ff8b8c5ce41a723fc6635a1d9587c307.

As the required version of po4a to support asciidoc attributes is v0.52
which was released january 2018, we will remove the attributes. Many
platforms we would like to support has not yet updated. We don't use the
icons for the asciidoc things now, so we don't need to do it in v5
either... There are lots of other things to spend the time on.