Gsg new workflow no netlist for pcbnew (#637)
* Getting Started: images in 4.1 "Using Eeschema"

Images correspond to KiCad master HEAD (dd9a00107 and before, 5.1-dev, in Linux Kubuntu/GTK+ 2.24).
MCU symbol has been updated. Also UI is a bit different, especially the symbol chooser.

* Getting Started: one more eeschema image for KiCad5.1

* Getting Started: 4.1.9: the clarification menu can't be summoned here
at all.
The note should be moved but this is a quick fix.

* Getting Started: 4.1.11: Ref is above, not inside the symbol.

* Getting Started: 4.1.19: UI button has changed.

Also fixed a spelling error.

* Getting Started: 4.1.19...20.

The component is already available. Let's add it explicitly to the project specific table,
not to either one. That way the global file isn't polluted (and there's no warning).

* Getting Started: 4.1.22: MCU must be flipped.

This way it's as close to the old design as possible.

* Getting Started: 4.1.29: In the new MCU symbol the ref and value are in the way.

* Getting Started: 4.1.32: There's a corner in the new design.

(And creating a corner was instructed nowhere AFAIK.)

* Getting Started: 4.1.43: the classic warning is changed.

* Getting Started: 4.1.46: There's no confirmation when annotating from scratch.

* Getting Started: Added notes for new workflows in eeshema.

Cvpcb is only one of the ways to assign footprints.
Netlist isn't needed.
This commit is a compromise between leaving everything as is
and moving completely to new workflows. The netlist can
be removed later or demoted to a note but it requires renumbering.

* Getting Started: Added direct pcb update from eeshcema workflow to Pcbnew section.